+++ udata/harvest/tests/test_base_backend.py
@@ -78,11 +78,10 @@ class BaseBackendTest:
         assert Dataset.objects.count() == nb_datasets
         for dataset in Dataset.objects():
             assert_equal_dates(dataset.last_modified, now)
-            assert dataset.extras['harvest:source_id'] == str(source.id)
-            assert dataset.extras['harvest:domain'] == source.domain
-            assert dataset.extras['harvest:remote_id'].startswith('fake-')
-            harvest_last_update = parse(dataset.extras['harvest:last_update'])
-            assert_equal_dates(harvest_last_update, now)
+            assert dataset.harvest.source_id == str(source.id)
+            assert dataset.harvest.domain == source.domain
+            assert dataset.harvest.remote_id.startswith('fake-')
+            assert_equal_dates(dataset.harvest.last_update, now)
 
     def test_has_feature_defaults(self):
         source = HarvestSourceFactory()
@@ -144,9 +143,9 @@ class BaseBackendTest:
         # no new datasets have been created
         assert len(datasets) == nb_datasets
         for dataset in datasets:
-            assert dataset.extras['harvest:source_id'] == str(source.id)
+            assert dataset.harvest.source_id == str(source.id)
             parsed = urlparse(source_url).netloc.split(':')[0]
-            assert parsed == dataset.extras['harvest:domain']
+            assert parsed == dataset.harvest.domain
 
     def test_dont_overwrite_last_modified(self, mocker):
         last_modified = faker.date_time_between(start_date='-30y', end_date='-1y')
@@ -158,8 +157,7 @@ class BaseBackendTest:
         dataset = Dataset.objects.first()
 
         assert dataset.last_modified == last_modified
-        harvest_last_update = parse(dataset.extras['harvest:last_update'])
-        assert_equal_dates(harvest_last_update, datetime.now())
+        assert_equal_dates(dataset.harvest.last_update, datetime.now())
 
     def test_dont_overwrite_last_modified_even_if_set_to_same(self, mocker):
         last_modified = faker.date_time_between(start_date='-30y', end_date='-1y')
@@ -172,8 +170,7 @@ class BaseBackendTest:
         dataset = Dataset.objects.first()
 
         assert dataset.last_modified == last_modified
-        harvest_last_update = parse(dataset.extras['harvest:last_update'])
-        assert_equal_dates(harvest_last_update, datetime.now())
+        assert_equal_dates(dataset.harvest.last_update, datetime.now())
 
     def test_autoarchive(self, app):
         nb_datasets = 3
@@ -183,21 +180,21 @@ class BaseBackendTest:
         # create a dangling dataset to be archived
         limit = app.config['HARVEST_AUTOARCHIVE_GRACE_DAYS']
         last_update = datetime.now() - timedelta(days=limit + 1)
-        dataset_arch = DatasetFactory(extras={
-            'harvest:domain': source.domain,
-            'harvest:source_id': str(source.id),
-            'harvest:remote_id': 'not-on-remote',
-            'harvest:last_update': last_update.isoformat(),
+        dataset_arch = DatasetFactory(harvest={
+                'domain': source.domain,
+                'source_id': str(source.id),
+                'remote_id': 'not-on-remote',
+                'last_update': last_update,
         })
 
         # create a dangling dataset that _won't_ be archived because of grace period
         limit = app.config['HARVEST_AUTOARCHIVE_GRACE_DAYS']
         last_update = datetime.now() - timedelta(days=limit - 1)
-        dataset_no_arch = DatasetFactory(extras={
-            'harvest:domain': source.domain,
-            'harvest:source_id': str(source.id),
-            'harvest:remote_id': 'not-on-remote-two',
-            'harvest:last_update': last_update.isoformat(),
+        dataset_no_arch = DatasetFactory(harvest={
+                'domain': source.domain,
+                'source_id': str(source.id),
+                'remote_id': 'not-on-remote-two',
+                'last_update': last_update,
         })
 
         job = backend.harvest()
@@ -213,26 +210,26 @@ class BaseBackendTest:
 
         dataset_arch.reload()
         assert dataset_arch.archived is not None
-        assert 'harvest:archived' in dataset_arch.extras
-        assert 'harvest:archived_at' in dataset_arch.extras
+        assert 'archived' in dataset_arch.harvest
+        assert 'archived_at' in dataset_arch.harvest
 
         dataset_no_arch.reload()
         assert dataset_no_arch.archived is None
-        assert 'harvest:archived' not in dataset_no_arch.extras
-        assert 'harvest:archived_at' not in dataset_no_arch.extras
+        assert 'archived' not in dataset_no_arch.harvest
+        assert 'archived_at' not in dataset_no_arch.harvest
 
         # test unarchive: archive manually then relaunch harvest
-        q = {'extras__harvest:remote_id': 'fake-1'}
+        q = {'harvest__remote_id': 'fake-1'}
         dataset = Dataset.objects.get(**q)
         dataset.archived = datetime.now()
-        dataset.extras['harvest:archived'] = 'not-on-remote'
-        dataset.extras['harvest:archived_at'] = datetime.now()
+        dataset.harvest.archived = 'not-on-remote'
+        dataset.harvest.archived_at = datetime.now()
         dataset.save()
         backend.harvest()
         dataset.reload()
         assert dataset.archived is None
-        assert 'harvest:archived' not in dataset.extras
-        assert 'harvest:archived_at' not in dataset.extras
+        assert 'archived' not in dataset.harvest
+        assert 'archived_at' not in dataset.harvest
 
     def test_harvest_datasets_get_deleted(self):
         nb_datasets = 3