+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 
 - Nothing yet
 
+## 5.0.1 (2022-11-14)
+
+- Fix resource harvest uri validation error [#2780](https://github.com/opendatateam/udata/pull/2780)
+
+## 5.0.0 (2022-11-14)
+
+- :warning: **Breaking change** Use dedicated dynamic harvest metadata for dataset and resources. A migration copies identifying fields from extras to this dedicated metadata field. Extras won't be used anymore for harvest-related information. udata-ckan, udata-ods and udata-front packages are impacted and should be upgraded accordingly [#2762](https://github.com/opendatateam/udata/pull/2762)
+
 ## 4.1.3 (2022-11-02)
 
 - Fix image URLs for suggest endpoints [#2761](https://github.com/opendatateam/udata/pull/2761)